requirements.txt
numpy
Flask
gunicorn
sklearn
pandas
PyYAML
evidently
dill
matplotlib
dask
kneed
joblib
-e .